refactor: handler 함수 분리 (#121)
